requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
sqlparse==0.5.3
urllib3==2.5.0
Django==4.2.16
djangorestframework==3.14.0
requests==2.31.0
python-decouple==3.8
psycopg2-binary==2.9.9
redis==4.6.0